Remove useless method. Add BasePhoneNumberExtractorConfiguration in python
