+++ Python/libraries/recognizers-sequence/recognizers_sequence/sequence/config/__init__.py
@@ -1,3 +1,4 @@
 from .url_configuration import *
 from .phone_number_configuration import *
 from .ip_configuration import *
+from .base_phone_number_configuration import *
